+++ docs/tutorials/grid5000/tuto_grid5000_docker.py
@@ -21,7 +21,11 @@ roles, networks = provider.init()
 registry_opts = dict(type="external", ip="docker-cache.grid5000.fr", port=80)
 
 d = en.Docker(
-    agent=roles["control"], bind_var_docker="/tmp/docker", registry_opts=registry_opts
+    agent=roles["control"],
+    bind_var_docker="/tmp/docker",
+    registry_opts=registry_opts,
+    # Optional credentials for docker hub
+    # credentials=dict(login="mylogin", password="mytoken"),
 )
 d.deploy()